requirements.txt
toolz==0.12.0
prettytable==3.9.0
streamlit==1.28.1
pandas==2.1.2
Pillow==10.0.1
